requirements.txt
fastapi
uvicorn
psycopg2-binary
requests
faster-whisper
python-dotenv